@@ -2,6 +2,7 @@ import { Component, Output, EventEmitter } from '@angular/core';
 import { DataStorageService } from '../../shared/data-storage.service';
 import { Response } from '@angular/http';
 import { AuthService } from '../../auth/auth.service';
+// import { HttpEvent, HttpEventType } from '@angular/common/http';
 
 @Component({
   selector: 'app-header',
@@ -16,8 +17,8 @@ export class AppHeaderComponent {
     onSaveData() {
         this.dataStorageService.storeRecipes()
             .subscribe(
-                (response: Response) => {
-                   // console.log(response);
+                (response) => {
+                   console.log(response);
                 }
             );
     }
