@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { Http, Response } from '@angular/http';
+import { HttpClient, HttpHeaders, HttpParams, HttpRequest } from '@angular/common/http';
 import 'rxjs/Rx';
 
 import { RecipeService } from '../recipes/recipe.service';
@@ -7,29 +7,42 @@ import { Recipe } from '../recipes/recipe.model';
 import { AuthService } from '../auth/auth.service';
 @Injectable()
 export class DataStorageService {
-    constructor(private http: Http,
+    constructor(private httpClient: HttpClient,
         private recipeService: RecipeService,
         private authService: AuthService) {}
 
     storeRecipes() {
         const token = this.authService.getToken();
+        const header = new HttpHeaders().set('Authorization', 'Bearer dfgdfh');
 
-        return this.http
-            .put('https://ng-recipebook-shoppinglist.firebaseio.com/recipes.json?auth=' + token, this.recipeService.getRecipes());
+    //     return this.httpClient
+    //         .put('https://ng-recipebook-shoppinglist.firebaseio.com/recipes.json', this.recipeService.getRecipes(),
+    //     { observe: 'body',
+    //     params: new HttpParams().set('auth', token)
+    //   // headers: header
+    // });
+        // const req = new HttpRequest('PUT', 'https://ng-recipebook-shoppinglist.firebaseio.com/recipes.json',
+        //                     this.recipeService.getRecipes(), {reportProgress: true, params: new HttpParams().set('auth', token)});
+        // return this.httpClient.request(req);
+
+        // With interceptor
+        const req = new HttpRequest('PUT', 'https://ng-recipebook-shoppinglist.firebaseio.com/recipes.json',
+        this.recipeService.getRecipes(), {reportProgress: true});
+return this.httpClient.request(req);
     }
 
     fetchRecipes() {
         const token = this.authService.getToken();
 
-        return this.http.get('https://ng-recipebook-shoppinglist.firebaseio.com/recipes.json?auth=' + token)
+        return this.httpClient.get<Recipe[]>('https://ng-recipebook-shoppinglist.firebaseio.com/recipes.json')
             .map(
-                (response: Response) => {
-                    const recipes: Recipe[] = response.json();
+                (recipes) => {
                     for (const recipe of recipes) {
                         if (!recipe['ingredients']) {
                             recipe['ingredients'] = [];
                         }
                     }
+                    // console.log(JSON.stringify(recipes));
                     return recipes;
                 }
             )
